+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Change Log for SD.Next
 
-## Update for 2025-07-28
+## Update for 2025-07-29
 
-### Highlights for 2025-07-28
+### Highlights for 2025-07-29
 
 Feature highlights include:  
 - **ModernUI** has quite some redesign which should make it more user friendly and easier to navigate plus several new UI themes!  
@@ -32,7 +32,7 @@ For details, see [ChangeLog](https://github.com/vladmandic/automatic/blob/master
 
 [ReadMe](https://github.com/vladmandic/automatic/blob/master/README.md) | [ChangeLog](https://github.com/vladmandic/automatic/blob/master/CHANGELOG.md) | [Docs](https://vladmandic.github.io/sdnext-docs/) | [WiKi](https://github.com/vladmandic/automatic/wiki) | [Discord](https://discord.com/invite/sd-next-federal-batch-inspectors-1101998836328697867)
 
-### Details for 2025-07-28
+### Details for 2025-07-29
 
 - **License**  
   - SD.Next [license](https://github.com/vladmandic/sdnext/blob/dev/LICENSE.txt) switched from **aGPL-v3.0** to **Apache-v2.0**  
@@ -169,6 +169,7 @@ For details, see [ChangeLog](https://github.com/vladmandic/automatic/blob/master
   - fix prompt encoding if prompts within batch have different segment counts  
   - fix detailer min/max size  
   - fix loopback script  
+  - fix yolo refresh models  
   - cleanup control infotext  
   - allow upscaling with models that have implicit VAE processing  
   - framepack improve offloading